[data-table] update test with card structure
